reuse diffTestData by moving into testutil
