+++ cli/cmd/edges_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/linkerd/linkerd2/controller/api/public"
+	"github.com/linkerd/linkerd2/testutil"
 )
 
 type edgesParamsExp struct {
@@ -118,5 +119,5 @@ func testEdgesCall(exp edgesParamsExp, t *testing.T) {
 	rows := edgesRespToRows(resp)
 	output := renderEdgeStats(rows, exp.options)
 
-	diffTestdata(t, exp.file, output)
+	testutil.DiffTestdata(t, exp.file, output, prettyDiff, updateFixtures, rejectPath)
 }
