+++ cli/cmd/install_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/linkerd/linkerd2/cli/flag"
 	charts "github.com/linkerd/linkerd2/pkg/charts/linkerd2"
 	"github.com/linkerd/linkerd2/pkg/tls"
+	"github.com/linkerd/linkerd2/testutil"
 	corev1 "k8s.io/api/core/v1"
 )
 
@@ -219,7 +220,7 @@ func TestRender(t *testing.T) {
 			if err := render(&buf, tc.values, ""); err != nil {
 				t.Fatalf("Failed to render templates: %v", err)
 			}
-			diffTestdata(t, tc.goldenFileName, buf.String())
+			testutil.DiffTestdata(t, tc.goldenFileName, buf.String(), prettyDiff, updateFixtures, rejectPath)
 		})
 	}
 }
