+++ cli/cmd/main_test.go
@@ -2,12 +2,8 @@ package cmd
 
 import (
 	"flag"
-	"io/ioutil"
 	"os"
-	"path/filepath"
 	"testing"
-
-	"github.com/sergi/go-diff/diffmatchpatch"
 )
 
 var (
@@ -29,59 +25,4 @@ func TestMain(m *testing.M) {
 	flag.StringVar(&rejectPath, "reject-path", "", "write results for failed tests to this path (path is relative to the test location)")
 	flag.Parse()
 	os.Exit(m.Run())
-}
-
-// readTestdata reads a file and returns the contents of that file as a string.
-func readTestdata(t *testing.T, fileName string) string {
-	file, err := os.Open(filepath.Join("testdata", fileName))
-	if err != nil {
-		t.Fatalf("Failed to open expected output file: %v", err)
-	}
-
-	fixture, err := ioutil.ReadAll(file)
-	if err != nil {
-		t.Fatalf("Failed to read expected output file: %v", err)
-	}
-
-	return string(fixture)
-}
-
-func writeTestdata(t *testing.T, fileName string, data []byte) {
-	p := filepath.Join("testdata", fileName)
-	if err := ioutil.WriteFile(p, data, 0644); err != nil {
-		t.Fatal(err)
-	}
-}
-
-func writeRejects(t *testing.T, origFileName string, data []byte) {
-	p := filepath.Join(rejectPath, origFileName+".rej")
-	if err := ioutil.WriteFile(p, data, 0644); err != nil {
-		t.Fatal(err)
-	}
-}
-
-// TODO: share this with integration tests
-func diffTestdata(t *testing.T, path, actual string) {
-	expected := readTestdata(t, path)
-	if actual == expected {
-		return
-	}
-	dmp := diffmatchpatch.New()
-	diffs := dmp.DiffMain(expected, actual, true)
-	diffs = dmp.DiffCleanupSemantic(diffs)
-	var diff string
-	if prettyDiff {
-		diff = dmp.DiffPrettyText(diffs)
-	} else {
-		diff = dmp.PatchToText(dmp.PatchMake(diffs))
-	}
-	t.Errorf("mismatch: %s\n%s", path, diff)
-
-	if updateFixtures {
-		writeTestdata(t, path, []byte(actual))
-	}
-
-	if rejectPath != "" {
-		writeRejects(t, path, []byte(actual))
-	}
 }
