+++ cli/cmd/routes_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/linkerd/linkerd2/controller/api/public"
+	"github.com/linkerd/linkerd2/testutil"
 )
 
 type routesParamsExp struct {
@@ -64,5 +65,5 @@ func testRoutesCall(exp routesParamsExp, t *testing.T) {
 		t.Fatalf("Unexpected error: %v", err)
 	}
 
-	diffTestdata(t, exp.file, output)
+	testutil.DiffTestdata(t, exp.file, output, prettyDiff, updateFixtures, rejectPath)
 }
