+++ cli/cmd/stat_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/linkerd/linkerd2/controller/api/public"
 	"github.com/linkerd/linkerd2/pkg/k8s"
+	"github.com/linkerd/linkerd2/testutil"
 )
 
 type paramsExp struct {
@@ -231,5 +232,5 @@ func testStatCall(exp paramsExp, resourceType string, t *testing.T) {
 	rows := respToRows(resp)
 	output := renderStatStats(rows, exp.options)
 
-	diffTestdata(t, exp.file, output)
+	testutil.DiffTestdata(t, exp.file, output, prettyDiff, updateFixtures, rejectPath)
 }
